Reorder the project to add git cmd source
