+++ proposals/gitcmd/source.go
@@ -1,0 +1,1 @@
+package gitcmd
